[SKIP]: Establish a messaging baseline
We'll start by establishing a foundational set of controller-, view-,
and model-based interactions. Authentication and
authorization are outside the scope of this article. For the sake of
this demonstration. Assume that visiting
<http://localhost:3000/sessions/new> presents visitors with a mechanism
to "log into" each `User` account.

```html+erb
<main class="flex flex-col h-screen">
  <section class="flex-1 flex flex-col overflow-y-scroll space-y-2">
    <%= render partial: "messages/message", collection: @messages.reverse %>
  </section>

  <aside>
    <%= form_with model: [ @user, @messages.new ] do |form| %>
      <%= form.label :content %>
      <%= form.rich_text_area :content, class: "h-24 overflow-y-scroll", autofocus: true %>

      <button>Send</button>
    <% end %>
  </aside>
</main>
```

`[autofocus]` has accessibility considerations worth discussing at
length. This example pre-supposes a lot about the rest of the
application, including an ideal user-experience that would result in a
user landing on the messages page with their cursor already in the
`<trix-editor>` field. We'll leave that journey as an exercise for the
reader.

```html+erb
<article id="<%= dom_id message %>" class="flex items-start">
  <%= image_tag message.sender.avatar, class: "rounded-full mx-4",
        width: 32, height: 32, alt: message.sender.name %>

  <div class="flex flex-col items-start space-y-1">
    <div class="rounded-xl bg-gray-100 p-2">
      <%= message.content %>
    </div>

    <time datetime="<%= message.created_at.iso8601 %>">
      <%= localize message.created_at, format: :long %>
    </time>
  </div>
</article>
```